Changed the regexp_replace to solve some issues
Simplified the body replacement part as "LANGUAGE" clause could be after the "AS" clause. The new regexp is much more simple and efficient.
A second regexp_replace had to be added to handle the case when "OUT/INOUT" parameters are specified. In that case, we have to remove the "RETURNS VOID" part added at the previous step in order to remain consistent with the parameters.